[CS]: Опциональные быстрые слоты из ЗП
- перенёс код квикслотов, сделал статики опциональными (при их отсутствии игра вылетать не будет)
- для работы адаптировал под ЧН метод UIDragDropReferenceList
- сразу сделал вызов квикслотов и ночного зрения так, чтобы они работали и например в машине
- добавил исправление на попытку запихнуть в статик квикслота предмет размером больше 1х1; т.е. можно привязать квики через консоль и абсолютно не волноваться о возможных вылетах, хоть гитару в слот запихивай (из OGSM, например)
- попытался исправить возможный вылет при перемещении предмета из статика квикслота в соседний статик квикслота (раньше вылетало, теперь работает без нареканий и меняет привязку квикслота, как и должно)

+++ SourcesAXR/xrGameCS/eatable_item.h
@@ -32,8 +32,8 @@ public:
 			void			UpdateUseAnim				(CActor* actor);
 			void			HideWeapon					();
 			void			StartAnimation				();
-			void SetRemainingUses(u8 value) { if (value > m_iConstPortions) return; m_iPortionsNum = value; };
-			u8 GetMaxUses() const { return m_iConstPortions; };
+			void			SetRemainingUses			(u32 value) { if (value > m_iConstPortions) return; m_iPortionsNum = value; };
+			u32				GetMaxUses					() const { return m_iConstPortions; };
 	virtual	void			UseBy						(CEntityAlive* npc);
 	virtual	bool			Empty						()						{return m_iPortionsNum==0;};
 	virtual	u32				Cost						()	const;
